Don't change network location if it's already set properly
If the current network location is correct, then don't change the
network location or display a notification.

To effect this change, switch from `scselect` to `networksetup` for
setting the network location (and for finding out the current network
location).

Also, check to make sure `networksetup` succeeds and display an error
notification if it doesn't.